requirements.txt
asgiref==3.3.1
captcha==0.3
cffi==1.14.4
cryptography==3.3.1
Django==3.1.3
django-appconf==1.0.4
django-ckeditor==6.0.0
django-filter==2.4.0
django-haystack==3.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.13
djangorestframework==3.12.2
importlib-metadata==3.1.0
Markdown==3.3.3
pilkit==2.0
Pillow==8.0.1
pycparser==2.20
PyMySQL==0.10.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
Whoosh==2.7.4
zipp==3.4.0
